requirements.txt
aws-cdk-lib==2.31.1
boto3==1.24.5
invoke==1.4.1
flake8==4.0.1
